rag-system/requirements.txt
# Core Dependencies
cryptography>=41.0.0
pycryptodome>=3.19.0

# Vector Database
qdrant-client>=1.7.0

# Embeddings & NLP
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0

# LLM Integration
langchain>=0.1.0
langchain-community>=0.0.10

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
python-pptx>=0.6.23
chardet>=5.2.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
